feat(packetloss): unify packet loss and traceroute monitoring interfaces
- Refactor UI components to merge packet loss and traceroute functionalities into a single interface.
- Update MTR integration to require root privileges and enhance fallback mechanisms for ICMP ping.
- Implement manual control for monitor startup behavior to prevent automatic execution on application launch.
- Improve user experience with clear navigation and seamless transitions between monitoring modes.
- Remove deprecated PacketLossTab component and streamline related functionalities.
