--- web/src/components/speedtest/PacketLossTab.tsx
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2024-2025, s0up and the autobrr contributors.
- * SPDX-License-Identifier: GPL-2.0-or-later
- */
-
-// Re-export the refactored PacketLossTab from its new location
-export { PacketLossTab } from "./packetloss/PacketLossTab";
