+++ web/src/components/speedtest/packetloss/components/EmptyStatePlaceholder.tsx
@@ -11,9 +11,126 @@ import {
   ExclamationTriangleIcon,
   ClockIcon,
   SignalIcon,
+  GlobeAltIcon,
+  PlayIcon,
 } from "@heroicons/react/24/outline";
 
-export const EmptyStatePlaceholder: React.FC = () => {
+interface EmptyStatePlaceholderProps {
+  mode?: "traceroute" | "monitors";
+  onHostSelect?: (host: string) => void;
+  onSwitchToMonitors?: () => void;
+}
+
+export const EmptyStatePlaceholder: React.FC<EmptyStatePlaceholderProps> = ({
+  mode = "monitors",
+  onHostSelect,
+  onSwitchToMonitors,
+}) => {
+  if (mode === "traceroute") {
+    return (
+      <motion.div
+        initial={{ opacity: 0, y: 20 }}
+        animate={{ opacity: 1, y: 0 }}
+        transition={{ duration: 0.5 }}
+        className="flex-1"
+      >
+        <div className="bg-gray-50/95 dark:bg-gray-850/95 rounded-xl p-6 shadow-lg border border-gray-200 dark:border-gray-800">
+          <div className="text-center mb-6">
+            <div className="w-16 h-16 bg-blue-500/10 rounded-full flex items-center justify-center mx-auto mb-4">
+              <GlobeAltIcon className="w-8 h-8 text-blue-600 dark:text-blue-400" />
+            </div>
+            <h2 className="text-xl font-semibold text-gray-900 dark:text-white mb-2">
+              Run a Traceroute
+            </h2>
+            <p className="text-gray-600 dark:text-gray-400 text-sm">
+              Discover every hop packets take to reach a destination
+            </p>
+          </div>
+
+          <div className="space-y-4">
+            <div className="p-4 bg-blue-500/5 border border-blue-500/20 rounded-lg">
+              <div className="flex items-start gap-3">
+                <PlayIcon className="w-5 h-5 text-blue-600 dark:text-blue-400 flex-shrink-0 mt-0.5" />
+                <div>
+                  <h3 className="text-blue-700 dark:text-blue-300 text-sm font-medium mb-1">
+                    Quick Start
+                  </h3>
+                  <p className="text-blue-700/80 dark:text-blue-300/80 text-xs mb-3">
+                    Enter any hostname or IP address to trace its route, or try
+                    these popular destinations:
+                  </p>
+                  {onHostSelect && (
+                    <div className="flex flex-wrap gap-2">
+                      <button
+                        onClick={() => onHostSelect("google.com")}
+                        className="px-2 py-1 bg-blue-500/20 hover:bg-blue-500/30 text-blue-700 dark:text-blue-300 rounded text-xs transition-colors"
+                      >
+                        google.com
+                      </button>
+                      <button
+                        onClick={() => onHostSelect("cloudflare.com")}
+                        className="px-2 py-1 bg-blue-500/20 hover:bg-blue-500/30 text-blue-700 dark:text-blue-300 rounded text-xs transition-colors"
+                      >
+                        cloudflare.com
+                      </button>
+                      <button
+                        onClick={() => onHostSelect("github.com")}
+                        className="px-2 py-1 bg-blue-500/20 hover:bg-blue-500/30 text-blue-700 dark:text-blue-300 rounded text-xs transition-colors"
+                      >
+                        github.com
+                      </button>
+                    </div>
+                  )}
+                </div>
+              </div>
+            </div>
+
+            <div className="grid grid-cols-1 md:grid-cols-2 gap-3">
+              <div className="flex items-start gap-3 p-3 bg-gray-200/30 dark:bg-gray-800/30 rounded-lg">
+                <div className="w-8 h-8 bg-emerald-500/10 rounded-lg flex items-center justify-center flex-shrink-0">
+                  <GlobeAltIcon className="w-4 h-4 text-emerald-600 dark:text-emerald-400" />
+                </div>
+                <div>
+                  <h4 className="text-gray-900 dark:text-white text-sm font-medium mb-1">
+                    Hop-by-Hop Analysis
+                  </h4>
+                  <p className="text-gray-600 dark:text-gray-400 text-xs">
+                    See each router in the path
+                  </p>
+                </div>
+              </div>
+
+              <div className="flex items-start gap-3 p-3 bg-gray-200/30 dark:bg-gray-800/30 rounded-lg">
+                <div className="w-8 h-8 bg-purple-500/10 rounded-lg flex items-center justify-center flex-shrink-0">
+                  <SignalIcon className="w-4 h-4 text-purple-600 dark:text-purple-400" />
+                </div>
+                <div>
+                  <h4 className="text-gray-900 dark:text-white text-sm font-medium mb-1">
+                    Create Monitor
+                  </h4>
+                  <p className="text-gray-600 dark:text-gray-400 text-xs">
+                    Convert results into a{" "}
+                    {onSwitchToMonitors ? (
+                      <button
+                        onClick={onSwitchToMonitors}
+                        className="text-purple-600 dark:text-purple-400 hover:text-purple-700 dark:hover:text-purple-300 underline"
+                      >
+                        scheduled monitor
+                      </button>
+                    ) : (
+                      "scheduled monitor"
+                    )}
+                  </p>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </motion.div>
+    );
+  }
+
+  // Monitors mode (default)
   return (
     <motion.div
       initial={{ opacity: 0, y: 20 }}
@@ -27,11 +144,10 @@ export const EmptyStatePlaceholder: React.FC = () => {
             <ChartBarIcon className="w-8 h-8 text-blue-600 dark:text-blue-400" />
           </div>
           <h3 className="text-xl font-semibold text-gray-900 dark:text-white mb-2">
-            Network Monitoring Dashboard
+            Packet Loss Monitoring
           </h3>
           <p className="text-gray-600 dark:text-gray-400 text-sm">
-            Select a monitor to view packet loss trends and network performance
-            data
+            Track packet loss percentage to specific hosts over time
           </p>
         </div>
 
@@ -43,10 +159,10 @@ export const EmptyStatePlaceholder: React.FC = () => {
               </div>
               <div>
                 <h4 className="text-gray-900 dark:text-white text-sm font-medium mb-1">
-                  MTR & ICMP Testing
+                  Scheduled Testing
                 </h4>
                 <p className="text-gray-600 dark:text-gray-400 text-xs">
-                  Advanced network diagnostics with automatic method selection
+                  Tests run automatically at your interval
                 </p>
               </div>
             </div>
@@ -57,10 +173,10 @@ export const EmptyStatePlaceholder: React.FC = () => {
               </div>
               <div>
                 <h4 className="text-gray-900 dark:text-white text-sm font-medium mb-1">
-                  Smart Alerts
+                  Packet Loss Alerts
                 </h4>
                 <p className="text-gray-600 dark:text-gray-400 text-xs">
-                  Threshold-based notifications for network issues
+                  Get notified when loss exceeds thresholds
                 </p>
               </div>
             </div>
@@ -71,10 +187,10 @@ export const EmptyStatePlaceholder: React.FC = () => {
               </div>
               <div>
                 <h4 className="text-gray-900 dark:text-white text-sm font-medium mb-1">
-                  Performance Trends
+                  Historical Charts
                 </h4>
                 <p className="text-gray-600 dark:text-gray-400 text-xs">
-                  Track network health over time
+                  View trends and patterns over time
                 </p>
               </div>
             </div>
@@ -85,10 +201,10 @@ export const EmptyStatePlaceholder: React.FC = () => {
               </div>
               <div>
                 <h4 className="text-gray-900 dark:text-white text-sm font-medium mb-1">
-                  Custom Intervals
+                  Smart Testing Method
                 </h4>
                 <p className="text-gray-600 dark:text-gray-400 text-xs">
-                  Set your own monitoring frequency
+                  MTR when available (requires root), ICMP ping otherwise
                 </p>
               </div>
             </div>
@@ -103,8 +219,8 @@ export const EmptyStatePlaceholder: React.FC = () => {
                 Get Started
               </h4>
               <p className="text-blue-700/80 dark:text-blue-300/80 text-xs">
-                Select a monitor to view detailed metrics and performance
-                charts.
+                Select a monitor to view packet loss history and real-time
+                status.
               </p>
             </div>
           </div>
